Fixed sender errors (power supply and off state).

+++ firmware_nousb/sender.h
@@ -1,4 +1,4 @@
 
 int send_iteration();
-int send_set_message(uint8_t *msg, int count);
+int send_set_message(uint8_t *msg, uint8_t count);
 int send_is_freely();
